Add unique index to investors SSN
This assumes that an individual investor can only appear in the
database once. I'm not really sure how this business model works
in this regard, but if the partner's get referral for users, then we'd
likely not want to pay more than once per individual investor.

Also adds #full_name convenience method.